session_rust::cylinder

Struct Cylinder

pub struct Cylinder {
    pub guid: String,
    pub name: String,
    pub radius: f64,
    pub line: Line,
    pub mesh: Mesh,
    pub xform: Xform,
}

A cylinder geometry defined by a line and radius.

The cylinder is generated as a 10-sided cylinder mesh that is oriented along the line direction and scaled to match the line length and specified radius.

Fields
guid: Stringname: Stringradius: f64line: Linemesh: Meshxform: Xform
Implementations

impl Cylinder

pub fn new(line: Line, radius: f64) -> Self

Creates a new Cylinder from a line and radius.

Arguments

line - The centerline of the cylinder

radius - The radius of the cylinder

Returns

A new Cylinder with a generated 10-sided cylinder mesh

pub fn transform(&mut self)

pub fn transformed(&self) -> Self

pub fn jsondump(&self) -> Result<String, Box<dyn Error>>

Serializes the Cylinder to a JSON string.

pub fn jsonload(json_data: &str) -> Result<Self, Box<dyn Error>>

Deserializes a Cylinder from a JSON string.

pub fn to_json(&self, filepath: &str) -> Result<(), Box<dyn Error>>

Serializes the Cylinder to a JSON file.

pub fn from_json(filepath: &str) -> Result<Self, Box<dyn Error>>

Deserializes a Cylinder from a JSON file.