session_rust::graph

Struct Graph

pub struct Graph {
    pub guid: String,
    pub name: String,
    pub vertex_count: i32,
    pub edge_count: i32,
    pub edges: HashMap<String, HashMap<String, Edge>>,
    /* private fields */
}

Fields
guid: Stringname: Stringvertex_count: i32edge_count: i32edges: HashMap<String, HashMap<String, Edge>>
Implementations

impl Graph

pub fn new(name: &str) -> Self

Creates a new, empty Graph with a specific name.

pub fn has_node(&self, key: &str) -> bool

Checks if a node exists in the graph.

pub fn add_node(&mut self, key: &str, attribute: &str) -> String

Adds a node to the graph.

pub fn add_edge( &mut self, u: &str, v: &str, attribute: &str, ) -> (String, String)

Adds an edge between u and v.

pub fn has_edge(&self, edge: (&str, &str)) -> bool

Checks if an edge exists in the graph.

pub fn number_of_vertices(&self) -> usize

Gets the number of vertices in the graph.

pub fn number_of_edges(&self) -> usize

Gets the number of edges in the graph.

pub fn get_vertices(&self) -> Vec<Vertex>

Gets all vertices in the graph.

pub fn get_edges(&self) -> Vec<(String, String)>

Gets all edges in the graph as tuples of vertex names.

pub fn neighbors(&self, node: &str) -> Vec<String>

Gets all neighbors of a node.

pub fn get_neighbors(&self, node: &str) -> Vec<String>

Gets all neighbors of a node (API compatibility method).

pub fn remove_node(&mut self, key: &str)

Removes a node and all its edges from the graph.

pub fn remove_edge(&mut self, edge: (&str, &str))

Removes an edge from the graph.

pub fn clear(&mut self)

Removes all vertices and edges from the graph.

pub fn node_attribute( &mut self, node: &str, value: Option<&str>, ) -> Option<String>

Get or set node attribute.

pub fn jsondump(&self) -> Result<String, Error>

Serializes the Graph to a JSON string.

pub fn jsonload(json_data: &str) -> Result<Self, Error>

Deserializes a Graph from a JSON string.

pub fn to_json(&self, filepath: &str) -> Result<(), Box<dyn Error>>

Serializes the Graph to a JSON file.

pub fn from_json(filepath: &str) -> Result<Self, Box<dyn Error>>

Deserializes a Graph from a JSON file.

pub fn edge_attribute( &mut self, u: &str, v: &str, value: Option<&str>, ) -> Option<String>

Get or set edge attribute.