session_rust::polyline

Struct Polyline

pub struct Polyline {
    pub guid: String,
    pub name: String,
    pub points: Vec<Point>,
    pub plane: Plane,
    pub width: f64,
    pub linecolor: Color,
    pub xform: Xform,
}

A polyline defined by a collection of points with an associated plane.

Fields
guid: Stringname: Stringpoints: Vec<Point>plane: Planewidth: f64linecolor: Colorxform: Xform
Implementations

impl Polyline

pub fn new(points: Vec<Point>) -> Self

Creates a new Polyline with default guid and name.

Arguments

points - The collection of points.

pub fn len(&self) -> usize

Returns the number of points in the polyline.

pub fn is_empty(&self) -> bool

Returns true if the polyline has no points.

pub fn segment_count(&self) -> usize

Returns the number of segments in the polyline. A polyline with n points has n-1 segments.

pub fn length(&self) -> f64

Calculates the total length of the polyline.

pub fn get_point(&self, index: usize) -> Option<&Point>

Returns a reference to the point at the given index.

pub fn get_point_mut(&mut self, index: usize) -> Option<&mut Point>

Returns a mutable reference to the point at the given index.

pub fn add_point(&mut self, point: Point)

Adds a point to the end of the polyline.

pub fn insert_point(&mut self, index: usize, point: Point)

Inserts a point at the specified index.

pub fn remove_point(&mut self, index: usize) -> Option<Point>

Removes and returns the point at the specified index.

pub fn reverse(&mut self)

Reverses the order of points in the polyline.

pub fn reversed(&self) -> Self

Returns a new polyline with reversed point order.

pub fn transform(&mut self)

pub fn transformed(&self) -> Self

pub fn jsondump(&self) -> Result<String, Box<dyn Error>>

Serializes the Polyline to a JSON string.

pub fn jsonload(json_data: &str) -> Result<Self, Box<dyn Error>>

Deserializes a Polyline from a JSON string.

pub fn to_json(&self, filepath: &str) -> Result<(), Box<dyn Error>>

Serializes the Polyline to a JSON file.

pub fn from_json(filepath: &str) -> Result<Self, Box<dyn Error>>

Deserializes a Polyline from a JSON file.

pub fn shift(&mut self, times: i32)

Shift polyline points by specified number of positions

pub fn length_squared(&self) -> f64

Calculate squared length of polyline (faster, no sqrt)

pub fn point_at_parameter(start: &Point, end: &Point, t: f64) -> Point

Get point at parameter t along a line segment (t=0 is start, t=1 is end)

pub fn closest_point_to_line( point: &Point, line_start: &Point, line_end: &Point, ) -> f64

Find closest point on line segment to given point, returns parameter t

pub fn line_line_overlap( line0_start: &Point, line0_end: &Point, line1_start: &Point, line1_end: &Point, ) -> Option<(Point, Point)>

Check if two line segments overlap and return the overlapping segment

pub fn line_line_average( line0_start: &Point, line0_end: &Point, line1_start: &Point, line1_end: &Point, ) -> (Point, Point)

Calculate average of two line segments

pub fn line_line_overlap_average( line0_start: &Point, line0_end: &Point, line1_start: &Point, line1_end: &Point, ) -> (Point, Point)

Calculate overlap average of two line segments

pub fn line_from_projected_points( line_start: &Point, line_end: &Point, points: &[Point], ) -> Option<(Point, Point)>

Create line from projected points onto a base line

pub fn closest_distance_and_point(&self, point: &Point) -> (f64, usize, Point)

Find closest distance and point from a point to this polyline

pub fn is_closed(&self) -> bool

Check if polyline is closed (first and last points are the same)

pub fn center(&self) -> Point

Calculate center point of polyline

pub fn center_vec(&self) -> Vector

Calculate center as vector

pub fn get_average_plane(&self) -> (Point, Vector, Vector, Vector)

Get average plane from polyline points

pub fn get_fast_plane(&self) -> (Point, Plane)

Get fast plane calculation from polyline

pub fn get_middle_line( line0_start: &Point, line0_end: &Point, line1_start: &Point, line1_end: &Point, ) -> (Point, Point)

Calculate middle line between two line segments

pub fn extend_line( line_start: &mut Point, line_end: &mut Point, distance0: f64, distance1: f64, )

Extend line segment by specified distances at both ends

pub fn scale_line(line_start: &mut Point, line_end: &mut Point, distance: f64)

Scale line segment inward by specified distance

pub fn extend_segment( &mut self, segment_id: usize, dist0: f64, dist1: f64, proportion0: f64, proportion1: f64, )

Extend polyline segment

pub fn extend_segment_equally_static( segment_start: &mut Point, segment_end: &mut Point, dist: f64, proportion: f64, )

Extend segment equally on both ends (static utility)

pub fn extend_segment_equally( &mut self, segment_id: usize, dist: f64, proportion: f64, )

Extend polyline segment equally

pub fn move_by(&mut self, direction: &Vector)

Move polyline by direction vector

pub fn is_clockwise(&self, _plane: &Plane) -> bool

Check if polyline is clockwise oriented

pub fn flip(&mut self)

Flip polyline direction (reverse point order)

pub fn get_convex_corners(&self) -> Vec<bool>

Get convex/concave corners of polyline

pub fn tween_two_polylines( polyline0: &Polyline, polyline1: &Polyline, weight: f64, ) -> Polyline

Interpolate between two polylines