Crate session_rust

Cross-language geometry library with Point, Color, and Vector types. Supports JSON serialization for interoperability between Rust, Python, and C++.

Re-exports

pub use arrow::Arrow;

pub use boundingbox::BoundingBox;

pub use bvh::BVH;

pub use color::Color;

pub use cylinder::Cylinder;

pub use edge::Edge;

pub use graph::Graph;

pub use line::Line;

pub use mesh::Mesh;

pub use obj::read_obj;

pub use obj::write_obj;

pub use objects::Objects;

pub use plane::Plane;

pub use point::Point;

pub use pointcloud::PointCloud;

pub use polyline::Polyline;

pub use quaternion::Quaternion;

pub use session::Geometry;

pub use session::Session;

pub use tolerance::Tolerance;

pub use tree::Tree;

pub use treenode::TreeNode;

pub use vector::Vector;

pub use vertex::Vertex;

pub use xform::Xform;

Modules

arrow

boundingbox

bvh

color

cylinder

edge

encoders

graph

intersection

line

mesh

obj

objects

plane

point

pointcloud

polyline

quaternion

session

tolerance

tree

treenode

vector

vertex

xform